New CommonService function: _request_termination()
This will tell the frontend that the service should be terminated. Use
this when the service has determined that it is unstable or has entered
an invalid state. There may be instances when the service may not be
able to recover from this by itself (eg. main thread stuck, situation
detected by a watchdog thread), so with this the frontend can step in
and ensure the service process is terminated.