Run callbacks in main service thread
Yeah... This is very much not obvious. But that thing we talked about,
callback functions called from a thread other than the subscribing
thread - turns out this is already covered via a special extra onion in
the subscription callback hell. The library callback functions (ie.
pika/stomp.py) don't actually call the callback function directly. cf.
the _on_message handler in stomp_transport.

The resulting code works. Reason for the crash in the SampleConsumer
was the passing through of pika-method and pika-properties, these can
just be discarded. There's now an opportunity to refactor out some
of this complexity.

I am also not convinced we ever need to invent subscription IDs, that
should already be sorted in the common_transport layer?