PikaTransport: use numeric subscription IDs internally (#85)
and only convert to strings at the pika API boundary.
Write the numeric subscription ID into message headers when passing
them to the subscribed callback functions, so they can be used to
ACK/NACK later.

The assertion preventing acknowledgements with multiple channels was
removed

This commit also adds some type annotations to the common transport
class, and make existing ones more precise. The SampleConsumer is
slightly modified to make it easier to use as a base for experiments.