Make every subscription run on a separate channel
And disable confirm mode on all channels.

This is so that transactions can be used on channels. So if a message is
to be send as part of a transaction then it can be sent on the relevant
channel, and if a message is sent outside of a transaction it can be
sent via the default channel.

The confirm mode currently does not have any effect as we are not
observing the callback. However, as confirm and transaction modes are
mutually exclusive remove the confirm mode from all subscription
channels. For the default channel we may eventually want to have one
in normal mode and one in confirm mode, but this is another issue
for another day.