PikaTransport: Enable Transactions (#92)
Make every subscription run on a separate channel so that
transactions can be used on channels. So if a message is
to be send as part of a transaction then it can be sent on the
relevant channel, and if a message is sent outside of a transaction
it can be sent via the default channel.

This disables confirm mode on all channels for now.
The confirm mode currently does not have any effect as we are not
observing the callback. However, confirm and transaction modes are
mutually exclusive, so transactions can't be used if we use confirm
mode on subscription channels. For the default channel we may
eventually want to have one in normal mode and one in confirm
mode, but this is another issue for another day.

Add optional subscription ID to transaction calls so that a transaction
can be associated with the correct channel where needed.