Enforce transaction semantics on RabbitMQ (#104)
If a transaction is running on an outgoing channel then all messages
going out on that channel must either match this transaction, or - if
they are outside of a transaction - must be committed immediately.
If a transaction is running, has uncommitted messages, and then a
message is sent outside of a transaction an error must be raised, as
RabbitMQ does not support this behaviour. Closes #97.

The only messages outgoing on a transacting, subscribing channel are
ACK/NACK and transacted sends. (Untransacted sends do not go through
this channel, so do not need to be considered.) In all three cases,
set a flag if there are uncommitted messages on the transaction to
detect the above case, so that an error can be raised.

This commit also introduces the use of bidirectional dictionaries from
the package 'bidict' to simplify tracking transactions.