Add common transport API for subscriptions to a temporary queue (#108)
This call creates a temporary queue, where the lifetime is tied to the specific
connection, and returns the subscription ID along with the queue name. This
name can then be given to some other process so that answers can be
received on that specific queue.

For more details see pull request #108 and the discussion in #105.