RMQ: support for delayed message exchanges
Delayed message exchanges specified with delayed_message_exchange
field in the rabbitmq configuration.

If no delayed message exchange is set, and a message is sent with
a delay, then raise a ValueError.
Otherwise send to the delayed message exchange if a delay is requested,
or to the default exchange if not.