Initial implementation of transport middleware
This enables user code to add middleware that can intercept or
monitor transport calls. Basic counter and timer middleware
are provided as examples.

Add PrometheusMiddleware that can be used to instrument downstream
code to collect prometheus metrics for exposing via a prometheus
endpoint.

Minimal example to add prometheus instrumentation is:

  import prometheus_client
  from workflows.transport.middleware.prometheus import PrometheusMiddleware

  instrument = PrometheusMiddleware(source="my-service")
  transport.add_middleware(instrument)
  prometheus_client.start_http_server(8000)

Add --no-strict-optional and --ignore-missing-imports to mypy
pre-commit-config to silence complaints from mypy not being able
to determine that self._pika_thread is not None in PikaTransport.