Add liveness probe to services
Add a --liveness option to zocalo.service which starts
a liveness endpoint on the port defined by --liveness-port
This sends a message to the frontend, which, if alive, will
return a message confirming it is alive. If confirmation of
aliveness is received, the endpoint will return a 200 status
code, else it returns 408 Request Timeout.

This is intended for use as a livenessProbe when running services
on Kubernetes. It should force a container restart in the event
that a pod is blocked indefinitely.