Add liveness probe to services (#150)
Add a --liveness option to zocalo.service which starts
a liveness endpoint on the port defined by --liveness-port
This sends a message to the service, which, if alive, will
return a message to the frontend confirming it is alive.
If confirmation of aliveness is received the endpoint will
return a 200 status code.

This is intended for use as a livenessProbe when running services
on Kubernetes. It should force a container restart in the event
that a pod is blocked indefinitely.