Sample pipethrough (#156)
New service to consume a message from a queue, and forward it to
another.

The sample producer and sample consumer services allow a message to be
sent between them. The sample pipethrough service produces a test
service which can be used to both receive and send messages in a recipe.

This is useful to test features that need some basic functionality over
multiple services, such as distributed tracing.